[secp256k1] ci: move test environment variable declaration to .cirrus.yml
Summary:
environment var moved:
    1. SECP256K1_TEST_ITERS (replaces TEST_ITERS)
    2. SECP256K1_BENCH_ITERS (replaces BENCH_ITERS)

This is a backport of secp256k1#1008
Depends on D19527

Backport note: we previously forgot to set the indirection to define SECP256K1_TEST_ITERS  in the cmake build script. This change has the added benefit that now the cmake and autotools build will both use the same number of test iterations in the cmake build script. This change has the added benefit that now the cmake and autotools build will both use the same number of test iterations. The cmake build will generally be faster on CI, because unit tests will  no longer default to 64 iterations.

Test Plan: push to github, wait for Github Actions to complete

Reviewers: #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D19528