secp256k1] build: Windows DLL additions
Summary:
Pass win32-dll to LT_INIT:

This is the recommended way to support building PE DLLs with modern
mingw toolchains and libtool.

> This option should be used if the package has been ported to build clean
> dlls on win32 platforms.
> If this macro is not used, libtool will assume that the package libraries
> are not dll clean and will build only static libraries on win32 hosts.

Add -no-undefined to libtool LDFLAGS:

Instruct libtool to not allow undefined symbols when linking a shared
library.

See:
https://autotools.io/libtool/windows.html
https://www.gnu.org/software/libtool/manual/libtool.html#LT_005fINIT
https://www.gnu.org/software/gnulib/manual/html_node/Libtool-and-Windows.html

This is a backport of core#1022

Test Plan: run the windows build on Cirrus-CI

Reviewers: #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D19469