[secp256k1] bump CI dockerfile
------

secp256k1#1583 & secp256k1#1690:

Bump GCC snapshot major version to 16
Clang should silently follow the main devel branch, but GCC needs to be bumped manually.

------

secp256k1#1720:

Corrects the spelling of "dpkg-dev" in a comment within the Dockerfile.

------

secp256k1#1756:

> Bump `docker/build-push-action` version
> See https://github.com/docker/build-push-action/releases.

https://github.com/bitcoin-core/secp256k1/pull/1756/changes/70ae177ca06b4d24c73c411d8524507d68aff45a

> Set `DEBIAN_FRONTEND=noninteractive`
> This suppresses `debconf: unable to initialize frontend: ...` warnings.

https://github.com/bitcoin-core/secp256k1/pull/1756/changes/f163c35897db9a4adc2a5b37926942817a4fdeb7

------

secp256k1#1758:

> ci: Drop workaround for Valgrind older than 3.20.0

This requires bumping the docker image from bookworm to trixie, which is the current `debian-stable`.
Otherwise we would need to duplicate the workaround for the cmake build to test valgrind with cmake.

------

This is a backport of secp256k1#1583, secp256k1#1690, secp256k1#1720, secp256k1#1756 (partial) and secp256k1#1758